+++ crogine/include/crogine/network/NetHost.hpp
@@ -103,7 +103,7 @@ namespace cro
         channels have higher priority, with 0 being highest.
         */
         template <typename T>
-        void broadcastPacket(uint32 id, const T& data, NetFlag flags, uint8 channel = 0);
+        void broadcastPacket(uint8 id, const T& data, NetFlag flags, uint8 channel = 0);
 
         /*!
         \brief Broadcasts the given stream of bytes to all connected clients
@@ -116,7 +116,7 @@ namespace cro
         \param channel Stream channel over which to send the data. Lower number
         channels have higher priority, with 0 being highest.
         */
-        void broadcastPacket(uint32 id, void* data, std::size_t size, NetFlag flags, uint8 channel = 0);
+        void broadcastPacket(uint8 id, void* data, std::size_t size, NetFlag flags, uint8 channel = 0);
 
         /*!
         \brief Sends a packet to the given peer if a connection is
@@ -135,7 +135,7 @@ namespace cro
         channels have higher priority, with 0 being highest.
         */
         template <typename T>
-        void sendPacket(const NetPeer& peer, uint32 id, const T& data, NetFlag flags, uint8 channel = 0);
+        void sendPacket(const NetPeer& peer, uint8 id, const T& data, NetFlag flags, uint8 channel = 0);
 
         /*!
         \brief Sends the given array of bytes out over the given peer if it
@@ -150,7 +150,7 @@ namespace cro
         \param channel Stream channel over which to send the data. Lower number
         channels have higher priority, with 0 being highest.
         */
-        void sendPacket(const NetPeer& peer, uint32 id, void* data, std::size_t size, NetFlag flags, uint8 channel = 0);
+        void sendPacket(const NetPeer& peer, uint8 id, void* data, std::size_t size, NetFlag flags, uint8 channel = 0);
 
 
         /*!
