+++ crogine/src/network/NetClient.cpp
@@ -202,7 +202,7 @@ bool NetClient::pollEvent(NetEvent& evt)
     return false;
 }
 
-void NetClient::sendPacket(uint32 id, void* data, std::size_t size, NetFlag flags, uint8 channel)
+void NetClient::sendPacket(uint8 id, void* data, std::size_t size, NetFlag flags, uint8 channel)
 {
     if (m_peer.m_peer)
     {
@@ -220,9 +220,9 @@ void NetClient::sendPacket(uint32 id, void* data, std::size_t size, NetFlag flag
             packetFlags |= ENET_PACKET_FLAG_UNRELIABLE_FRAGMENT | ENET_PACKET_FLAG_UNSEQUENCED;
         }
         
-        ENetPacket* packet = enet_packet_create(&id, sizeof(uint32), packetFlags);
-        enet_packet_resize(packet, sizeof(uint32) + size);
-        std::memcpy(&packet->data[sizeof(uint32)], data, size);
+        ENetPacket* packet = enet_packet_create(&id, sizeof(uint8), packetFlags);
+        enet_packet_resize(packet, sizeof(uint8) + size);
+        std::memcpy(&packet->data[sizeof(uint8)], data, size);
 
         enet_peer_send(m_peer.m_peer, channel, packet);
     }
